+++ orchestration/app/event_registry.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	corecontext "github.com/fluxplane/agentruntime/core/context"
+	coreconversation "github.com/fluxplane/agentruntime/core/conversation"
 	"github.com/fluxplane/agentruntime/core/event"
 	"github.com/fluxplane/agentruntime/core/operation"
 	"github.com/fluxplane/agentruntime/core/resource"
@@ -60,6 +61,9 @@ func defaultEventTypes() []event.Event {
 		operation.OperationFailed{},
 		operation.OperationRejected{},
 		operation.OperationCanceled{},
+		coreconversation.ItemsAppended{},
+		coreconversation.ContinuationStored{},
+		coreconversation.CompactionStored{},
 		corecontext.BlockRecorded{},
 		corecontext.BlockRemovedRecorded{},
 		corecontext.RenderCommitted{},